Rename files and move redux into components
