+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import * as serviceWorker from './serviceWorker';
-import App from './components/app/App';
+import App from './components/app/app.component';
 import { Provider } from 'react-redux'
 import store from './redux/store'
